# 8308118: Avoid multiarray allocations in AESCrypt.makeSessionKey