One of our services has a hot path with AES/GCM cipher reuse. The JDK code reinitializes the session key on that path, and [JDK-8308105](https://bugs.openjdk.org/browse/JDK-8308105) shows up prominently there.

Fixing [JDK-8308105](https://bugs.openjdk.org/browse/JDK-8308105) would take a while, as would likely require multiple patches in VM internals. Meanwhile, we can avoid the multiarray allocations in AESCrypt.makeSessionKey completely, reaping performance benefits. We can go even deeper: replace the multi-array with the flat array and drop `expandToSubKey` completely.

Example original profile is in the bug.

There are other things we can polish in that code, but experiments show those polishings have rather diminshed returns.

On new benchmark:

```
Benchmark       Mode  Cnt    Score   Error  Units

## Mac M1

# Before
AESReinit.test  avgt   15   873,842 ± 6,911  ns/op

# After
AESReinit.test  avgt   15   347,632 ± 8,764  ns/op  ; <--- 2.5x faster

## Xeon, c6.8xlarge

# Before
AESReinit.test  avgt   15  1524.307 ± 24.231 ns/op

# After
AESReinit.test  avgt   15   554.727 ± 12.876 ns/op  ; <--- 2.75x faster

## Graviton, m6g.4xlarge

# Before
AESReinit.test  avgt   15  1913.492 ± 23.489 ns/op

# After
AESReinit.test  avgt   15   639.701 ± 5.033  ns/op  ; <--- 2.99x faster
```

Additional testing:
 - [x] Benchmarks
 - [x] macos-aarch64-server-release, `jdk_security`
 - [x] linux-x86_64-server-fastdebug, `tier1 tier2 tier3`

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must be properly reviewed (1 review required, with at least 1 [Reviewer](https://openjdk.org/bylaws#reviewer))
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue

### Issue
 * [JDK-8308118](https://bugs.openjdk.org/browse/JDK-8308118): Avoid multiarray allocations in AESCrypt.makeSessionKey


### Reviewers
 * [Xue-Lei Andrew Fan](https://openjdk.org/census#xuelei) (@XueleiFan - **Reviewer**)
 * @schlosna (no known openjdk.org user name / role)


### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.org/jdk.git pull/13996/head:pull/13996` \
`$ git checkout pull/13996`

Update a local copy of the PR: \
`$ git checkout pull/13996` \
`$ git pull https://git.openjdk.org/jdk.git pull/13996/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 13996`

View PR using the GUI difftool: \
`$ git pr show -t 13996`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.org/jdk/pull/13996.diff">https://git.openjdk.org/jdk/pull/13996.diff</a>

</details>


### Webrev
[Link to Webrev Comment](https://git.openjdk.org/jdk/pull/13996#issuecomment-1548530219)